Deep audit: naming, implementation, and layering compliance
Deduplicate ~400 lines between Atomic and Streaming write APIs into
shared helpers (Kernel.File.Write+Shared.swift) with proper error
mapping. Remove dead Darwin/Glibc/WinSDK imports from L3 code.
Fix TOCTOU rename pattern, syncFile full/dataOnly distinction.

Naming compliance ([API-NAME-001], [API-NAME-002]):
- AffinityAccessor eliminated — static methods on Affinity
- ConditionAccessor → Channel, BroadcastAll → Broadcast
- arriveAndWait → arrive, arrivedCount → arrived
- waiterCount → waiters, writeRaw → write (type overload)
- currentState → state, Options booleans → Preservation + Ownership
- PATTERN-016 for compound methods needing Property-primitives

Implementation compliance ([IMPL-021], [PATTERN-017], [API-IMPL-005]):
- Barrier uses Synchronization<1>, Context state is enum
- Process.ID.rawValue → CustomStringConvertible
- Permissions(rawValue:) → .ownerReadWrite
- One type per file (Synchronization splits, Worker.Token, Write namespace)
- Typed throws on streaming fill closure

+++ Sources/Kernel/Kernel.Thread.Executors.swift
@@ -69,7 +69,7 @@ extension Kernel.Thread.Executors {
     /// - Returns: The next executor in the round-robin sequence.
     public func next() -> Kernel.Thread.Executor {
         let index = counter.wrappingAdd(1, ordering: .relaxed).oldValue
-        return executors[Int(index % UInt64(executors.count))]
+        return executors[Int(index) % executors.count]
     }
 
     /// Get a specific executor by index.
